refactor: clean up headers
