+++ src/include/btreeutils.h
@@ -4,6 +4,9 @@
 
 #include <linux/btree.h>
 
+#define LONG_PER_U64 (64 / BITS_PER_LONG)
+#define MAX_KEYLEN	(2 * LONG_PER_U64)
+
 struct btree {
 	struct btree_head *head;
 };
