+++ src/include/skiplist.c
@@ -10,12 +10,6 @@
 
 #include "skiplist.h"
 
-#define HEAD_KEY ((sector_t)0)
-#define HEAD_DATA NULL
-#define TAIL_KEY ((sector_t)U64_MAX)
-#define TAIL_DATA ((sector_t)0)
-#define MAX_LVL 20
-
 static void free_node_full(struct skiplist_node *node)
 {
 	struct skiplist_node *temp;
